Mode Dataset Resolution to Prepare Data
