+++ aiu_fms_testing_utils/testing/dpp/prepare_data.py
@@ -1,11 +1,17 @@
 import itertools
 import json
 import random
+
+from huggingface_hub import hf_hub_download
 from aiu_fms_testing_utils.testing.dpp.prepare_programs import get_programs_to_test
 from aiu_fms_testing_utils.testing.dpp.program_models import (
     PreparedInputs,
     ProgramInfo,
     ValidPrompt,
+)
+from aiu_fms_testing_utils.testing.dpp.run_drive_paged_programs import (
+    RAG_FACTOID_DATASET,
+    SHARE_GPT_DATASET,
 )
 from aiu_fms_testing_utils.utils.aiu_setup import dprint, local_rank
 from fms.utils.generation import pad_input_ids
@@ -307,3 +313,41 @@ def prepare_test_prompts(
         custom_shape=custom_shape,
         pad_multiple=PAD_MULTIPLE,
     )
+
+
+def resolve_dataset_path(dataset_path: str) -> tuple[str, str]:
+    """Resolves the dataset type and local path based on the provided dataset_path.
+
+    Args:
+        dataset_path: A string indicating the dataset to use. Supported values are:
+                      - "sharegpt": Uses the ShareGPT dataset from HuggingFace.
+                      - "rag_factoid": Uses the RAG Factoid dataset from HuggingFace.
+                      - Any other string is considered a custom dataset path.
+
+    Returns:
+        A tuple containing:
+            - dataset_type: A string indicating the type of dataset ("sharegpt", "rag_factoid", or "custom").
+            - local_dataset_path: The local file path to the dataset."""
+
+    if dataset_path == "sharegpt":
+        dataset_type = "sharegpt"
+        # Fetch from HuggingFace
+        local_dataset_path = hf_hub_download(
+            repo_id=SHARE_GPT_DATASET[0],
+            filename=SHARE_GPT_DATASET[1],
+            repo_type="dataset",
+        )
+    elif dataset_path == "rag_factoid":
+        dataset_type = "rag_factoid"
+        local_dataset_path = hf_hub_download(
+            repo_id=RAG_FACTOID_DATASET[0],
+            filename=RAG_FACTOID_DATASET[1],
+            repo_type="dataset",
+        )
+    elif dataset_path is None:
+        dataset_type = "custom"
+        local_dataset_path = dataset_path
+    else:
+        raise ValueError(f"Unsupported dataset_path: {dataset_path}")
+
+    return dataset_type, local_dataset_path
